requirements.txt
flask==3.0.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv
